refactor(storage): re-structure files
