+++ src/context/Services/storage/actions.ts
@@ -1,0 +1,15 @@
+import {
+  Actions, State, IsWhitelistedProviderPayload,
+} from './interfaces'
+
+const actions: Actions = {
+  SET_IS_WHITELISTED_PROVIDER: (
+    state: State,
+    { isWhitelistedProvider }: IsWhitelistedProviderPayload,
+  ): State => ({
+    ...state,
+    isWhitelistedProvider,
+  }),
+}
+
+export default actions
